+++ contracts/library/MerkleTreeLib.sol
@@ -43,15 +43,15 @@ library MerkleTreeLib {
     uint8 internal constant MERGE_ZEROS = 2;
     uint8 internal constant MAX_TREE_LEVEL = 255;
 
-    function isRight(bytes32 _hash, uint height) internal pure returns (bool) {
+    function isRight(bytes32 _hash, uint256 height) internal pure returns (bool) {
         return getBit(_hash, height);
     }
 
-    function setBit(bytes32 bitmap, uint index) internal pure returns (bytes32) {
+    function setBit(bytes32 bitmap, uint256 index) internal pure returns (bytes32) {
         return bytes32(uint256(bitmap) | (1 << (index & 0xff)));
     }
 
-    function getBit(uint256 bitmap, uint index) internal pure returns (bool) {
+    function getBit(uint256 bitmap, uint256 index) internal pure returns (bool) {
         bool result;
         assembly {
             let shifted := shl(index, 1)
@@ -62,7 +62,7 @@ library MerkleTreeLib {
         return result;
     }
 
-    function getBit(bytes32 bitmap, uint index) internal pure returns (bool) {
+    function getBit(bytes32 bitmap, uint256 index) internal pure returns (bool) {
         bool result;
         assembly {
             let shifted := shl(index, 1)
@@ -73,15 +73,15 @@ library MerkleTreeLib {
         return result;
     }
 
-    function clearBit(bytes32 bitmap, uint index) internal pure returns (bytes32) {
+    function clearBit(bytes32 bitmap, uint256 index) internal pure returns (bytes32) {
         return bytes32(uint256(bitmap) & (~(1 << index)));
     }
 
-    function copyBits(bytes32 bitmap, uint index) internal pure returns (bytes32) {
+    function copyBits(bytes32 bitmap, uint256 index) internal pure returns (bytes32) {
         return bytes32((uint256(bitmap) << index) >> index);
     }
 
-    function parentPath(bytes32 path, uint height) internal pure returns (bytes32) {
+    function parentPath(bytes32 path, uint256 height) internal pure returns (bytes32) {
         if (height == 255) {
             return bytes32(0);
         }
@@ -121,7 +121,7 @@ library MerkleTreeLib {
         bytes32[] calldata siblings
     ) internal pure returns (bool) {
         bytes32 parent_path;
-        uint iReverse;
+        uint256 iReverse;
         uint8 n;
         bool _isRight;
         MergeValueType mergeType;
@@ -149,7 +149,7 @@ library MerkleTreeLib {
             currentZeroBits = firstZeroBits;
         }
 
-        for (uint i = startIndex; ; ) {
+        for (uint256 i = startIndex; ; ) {
             unchecked {
                 iReverse = MAX_TREE_LEVEL - i;
             }
